Renamed Billable trait as per guidelines, fixes #221
